backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
sqlalchemy>=2.0
alembic>=1.14
psycopg2-binary>=2.9
pydantic>=2.10
pydantic-settings>=2.6
httpx>=0.28
feedparser>=6.0.11
bleach>=6.2
openai>=1.60
python-dateutil>=2.9

